+++ apps/consent-ui/src/components/common/Header/Header.module.scss
@@ -8,6 +8,10 @@
 	z-index: 1;
 }
 
+.logoLink {
+	display: flex;
+}
+
 .headerBar {
 	@apply grid grid-cols-2;
 	padding: 0.5rem 1rem;
